requirements.txt
flask
flask-sqlalchemy
authlib
python-dotenv
